backend/requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.1
flake8==3.6.0
mccabe==0.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.9
